chore: rename swr folder to core
